+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## UPCOMING
 
+## `v2021_02_18`
+* `Only run those tests that has config changes`
+
 ## `v2021_02_17`
 * `XCBeautify is now hardcoded`